+++ src/containers/index.ts
@@ -4,3 +4,4 @@ export * from './List';
 export { default as Tags } from './Tags';
 export { default as Search } from './Search';
 export { default as Todo } from './Todo';
+export { default as CreateTodo } from './CreateTodo';
