+++ src/libs/constant.ts
@@ -3,3 +3,4 @@ export const SECONDARY_COLOR_GRAY_1 = '#eff4fb';
 export const SECONDARY_COLOR_GRAY_2 = '#d6dae5';
 export const SECONDARY_COLOR_DARK = '#232323';
 export const SECONDARY_COLOR_WHITE = '#fff';
+export const TITLE_MAX_LENGTH = 30;
